Separate data struct graph in a specific file
