+++ src/all_cycles_in_graph/c++/detect_cycles.cpp
@@ -1,19 +1,6 @@
 #include "detect_cycles.h"
-#include <iostream>
 
-
-void Graph::add_arc (int source, int destination){
-    add_node(source);
-    add_node(destination);
-    arcs[source].push_back(destination);
-}
-
-void Graph::add_node (int node){
-    nodes.insert(node);
-}
-
-
-void DetectCycle::run (Graph graph){
+void DetectCycle::run(Graph graph){
     colors.clear();
     for(auto node : graph.nodes)
         colors[node] = WHITE;
@@ -25,7 +12,7 @@ void DetectCycle::run (Graph graph){
    }
 }
 
-void DetectCycle::dfs (Graph graph, int source){
+void DetectCycle::dfs(Graph graph, int source){
     colors[source] = GREY;
     for(auto destination: graph.arcs[source]){
         path[source] = destination;
