+++ src/all_cycles_in_graph/c++/detect_cycles.h
@@ -2,19 +2,11 @@
 #define DETECT_CYCLES_H
 
 #include <list>
-#include <set>
 #include <map>
-#include <math.h>
+
+#include "graph.h"
 
 using namespace std;
-
-class Graph{
-    public:
-        map<int, list<int> > arcs;
-        set<int> nodes;
-        void add_arc (int, int);
-        void add_node (int);
-};
 
 class DetectCycle{
     map<int, int> path;
