+++ src/all_cycles_in_graph/c++/tests/tests_detect_cycles.cpp
@@ -1,4 +1,5 @@
 #include <gtest/gtest.h>
+#include "../graph.h"
 #include "../detect_cycles.h"
 
 TEST(DetectCyclesTest, GraphWithoutArcs) {
